Move all halo related code from the driver to dedicated file
